requirements.txt
aioice==0.10.1
aiortc==1.11.0
altair==5.5.0
attrs==25.3.0
av==14.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
dnspython==2.7.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
huggingface-hub==0.30.2
idna==3.10
ifaddr==0.2.0
importlib_resources==6.5.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.4
more-itertools==10.6.0
mpmath==1.3.0
narwhals==1.34.1
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pip==25.0.1
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
setuptools==58.1.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
soundfile==0.13.1
streamlit==1.44.1
streamlit-audiorecorder==0.0.6
streamlit-webrtc==0.62.4
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
zipp==3.21.0